overwrite the old challenges because weird
